Use the fridge when building
In order to use the fridge along the oven a small
refactoring is needed. Basically, the fridge should
be able to run some steps before/after the oven takes place.
That means that we need a new API to cook a specific
recipe and not only a list of recipes plus its deps.
Now, whenever an unfreezing process fails (if the
--use-binary option is set) the recipe is truly built
and if succeeds it is uploaded.
All code that uses the old oven is alos ported