Avoid a circular dependency with freetype/harfbuzz
Freetype can be built with harfbuzz support and harfbuzz
can be built using freetype. On a clean environment
cerbero will first build freetype because of the
dependency tree, so far so good. Then, harfbuzz will be built
with freetype support. If later, freetype needs to be rebuilt
but harfbuzz is already installed, freetype will depend on it
and make a circular dependency. Depending on the linker it might
work, but on OSX we get an assert because of this, a counter
exceeds its limit and anything that needs freetype/harfbuzz or any
of its dependencies can't be linked