Several changes for Mingw >5.3 compatibility
On gstreamer:
------------
- Also check for winpthread library

In new versions of mingw clock_gettime and other time functions
are not longer part of the standard library, they have been
moved to wipthread library instead
This patch applies a patch into gstreamer's configure.ac in
order to check if clock_gettime is being provided by winpthread

On libtheora:
------------
- Do not redefine rint function for mingw

All versions of mingw correctly define 'rint' function in
math.h
However, only newer versions complain about the function
redefinition when the -O option is greater than zero
(the redefinition of having a const version for something
that was previously non-const)
Cleanest solution is do not redefine this function for
GNU-C family of compiler and use the compiler's
version instead.

- Remove CR/LF in .de definition files.

Having CR/LF in .def files makes libtool not detect
the fist line with the 'EXPORTS' tag
(because it is actually EXPORTS\r), and at the end a new .def
file with 'EXPORTS' duplicated is generated.
Having 'EXPORTS' twice is not supported by newer mingw versions.
The most reasonable patch (instead of patching libtool) is
simply having the .def files properly formatted with LF only