cerbero: Fix recipe and bootstrap unpack errors on Windows
This regressed after https://gitlab.freedesktop.org/gstreamer/cerbero/merge_requests/197

tar is much faster than tarfile, but the tar that use on Windows is
an ancient MSYS tar with many bugs that make it impossible to use it.
So, we use tar only if we're not running on Windows.