Decrease possibility of upload contention using fridge
Upload first the checksum to use it as a lock resource
so that the rest don't upload anything *if the checksum
is the same*. If the checksum changes, both the checksum
and the package will be overwritten. That's ok, though.
As long as the recipe is the same, packages are equally
usable.