libnice.recipe: Port interface discovery APIs to UWP
The following APIs are not allowed: `GetIfTable` `GetBestInterface`
`GetIpAddrTable`, so replace with equivalents. In the process, also
unbreak the service discovery APIs. Upstream MR has more details:
https://gitlab.freedesktop.org/libnice/libnice/-/merge_requests/154

WACK wasn't reporting an error for `GetIpAddrTable`, but the
documentation says it is only available on desktop:
https://docs.microsoft.com/en-us/windows/win32/api/iphlpapi/nf-iphlpapi-getipaddrtable#requirements

Part-of: <https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/546>