Add checksum of deps to make it safer
Even though sometimes the recipe can be the same,
the resulting artifact can be different. An example
is when the configure for a given recipe does something
different in case of different versions of its dependencies.
e.g. flu-codec-sdk and the files it includes depends
on the version of glib.

This way, we make the checksum more robust.