Recipe module for explicit class name resolution
This allows to infer details about the recipe using introspection
outside of the recipe folder more accurately. For example, before
this change using the function inspect.getsourceline to get the path
of class in the recipes folder from it's class object; but making this
call from a function outside of the recipes  folder, it will return
the class reference custom.GStreamer. The problem is if there is
another custom.py in the scope you are calling getsourcelines from.
The function getsourceline will do a best guess and may not reference
the correct custom.py. So, by turning the recipes folder into a
module we get the full class name resolution of
recipes.custom.GStreamer; instead of custom.GStreamer.This avoids
any confusion to which custom.py we are referring to.