Only update installed files when files are actually installed
Before this change, even if the package was already extracted,
update_installed_files was called because the one that knew
whether fridge steps were necessary to run was fridge. Now,
the oven also considers this and will only update the files
when no fridge package extraction is done OR when an
install or post_install step is actually run.