Ensure fridge's package integrity
Previously, the remote hash was only checked when local file
already existed to avoid doing networking as much as possible
relying on local cache. However, this has been proven to be
cause of issue. Thus, let's be explicit and pedantic and use
the checksum for files properly.

From now on, if the remote checksum does not match the file,
we assume it's a corrupted file. So, the recipe will be built
from scratch and re-uploaded.

Corrupted files are possible because of error when uploading
the files through FTP.