source: git: Avoid cloning common sources multiple times
In some cases, multiple recipes in a cbc may point to the same
git repo. The previous behaviour was to fetch for if recipe. In
this commit, when the repo_dir among multiple recipes is the
same, then only one of them will be fetched and the other
ignored.

There are two cases to ignore fetching a recipe:
 - The recipe was already fetch according the recipe status
 - There is other similar recipe fetching

An asyncio condition is used per group of recipes pointing to a
common git source. It is used to prevent the fetch step done if
there is an-in-progress fetch (for its source); so if the
in-progress feth fails, the others waiting are not considered
fetched in the cache file.

Issue: OCP-1914