Fix CMake not included in fetch step if needed
Commit 3208cb85fefe31564cb9249bff0cfad1aa4ce772 fixed one issue
but didn't fully covered all cases. It turns out that the build
tools (only CMake as of now) need to be checked after the bootstrap
has done the apt/yum installation.

Previous commit moved the logic to the `start` step, forgetting that
`fetch_recipes` also needs to consider the same. We could do that with
a decorator for both `start` and `fetch_recipes`, but I do not think
it is worth it for what we need to do right now. So, let's keep it
simple and add it in both sides.