Use host and target info for env instead of compiler  version
This is a trade-off to favor simplicity and speed
over safety. The reasoning behind this is that if
we do not need the compiler version, we can use
fridge before bootstrapping with consistency. Take
into account that bootstrap process may install a
different or newer compiler version.

Thus, this way we can use check-binaries-remote
before doing anything at all without needing to
bootstrap first. In any case, we add some nice
info for both host and target to ensure minimum
ABI compatibility.