Fix different checksum depending on deps order
For some reason, fluendo-sdk recipe includes duplicated
deps. Also, they might be included in different order.
So, let's do some defensive programming in all the lists
considered for the checksum of the recipe to ensure
consistency across different runs.

This was already done for the parents checksum, but I
wrongly assumed that list_deps() returned a sorted list.