Ensure fridge packages are only fetched when available
This tackles the scenario where fetch_binary is attempted
even if no remote binary exists. In this case, the % shown
for both fetch_binary and fetch would be the same, which
is quite confusing. We need to do some extra work to address
this though: we check async and concurrently before ever
attempting to run fetch_binary whether the package exists
in the remote or not.