linux.config: Fix gstreamer cross compilation
GStreamer configure.ac uses a test program for check if unaligned access is
possible. The safer or more conservative value for this is zero because
it's been used for set the macro `GST_HAVE_UNALIGNED_ACCESS` which chooses
between faster and slower implementations for read/swap integers from a
memory buffer.

Issue: OCP_3624