Strip: using generic utils.get_build_env() method to fix cross win issues
Warning: This commit fixes the Strip class, but not its usage.

Warning from @ylatuya Andoni Morales Alastruey:

Strip is a wrapper for the strip tool that should only be used for binaries built with gcc,
it should never be used for msvc-built binaries since it usually corrupts them.
In the case of msvc builds, debug symbols are stored in a separate .pdb file
and they don't require stripping.
Only recipes using_msvc=false should strip binaries.